fix(src/cdk): index reset
